Add redirects for edit/delete routes when login user != book/author/genre creator

+++ views.py
@@ -1,5 +1,5 @@
 from models import *
-from flask import Flask, jsonify, request, url_for, render_template, flash, redirect, make_response
+from flask import Flask, jsonify, request, url_for, render_template, flash, redirect, make_response, Markup
 from flask import session as login_session
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker
@@ -20,7 +20,8 @@ import requests
 
 app = Flask(__name__)
 
-engine = create_engine('sqlite:///books-02.db')
+# engine = create_engine('sqlite:///books-02.db')
+engine = create_engine('sqlite:///test-it.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -467,6 +468,9 @@ def edit_book(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   book = session.query(Book).filter_by(id = id).one()
+  # if book.user_id != login_session['user_id']:
+  #   flash(Markup("Sorry, you can only edit books that you created. <a href='/catalog/book/create'>Create your own book</a> if you're up for it!"))
+  #   return redirect(url_for('catalog_book', id = id))
   if request.method == 'POST':
     if request.form['title']:
       book.title = request.form['title']
@@ -504,6 +508,9 @@ def delete_book(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   book = session.query(Book).filter_by(id = id).one()
+  if book.user_id != login_session['user_id']:
+    flash(Markup("Sorry, you can only edit books that you created. <a href='/catalog/book/create'>Create your own book</a> if you're up for it!"))
+    return redirect(url_for('catalog_book', id = id))
   if request.method == 'POST':
     session.delete(book)
     session.commit()
@@ -555,6 +562,8 @@ def edit_genre(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   genre = session.query(Genre).filter_by(id = id).one()
+  if genre.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit genres that you created. <a href='/catalog/genre/create'>Create your own genre</a> if you're up for it!")
   if request.method == 'POST':
     if request.form['type']:
       genre.type = request.form['type']
@@ -581,6 +590,8 @@ def delete_genre(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   genre = session.query(Genre).filter_by(id = id).one()
+  if genre.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit genres that you created. <a href='/catalog/genre/create'>Create your own genre</a> if you're up for it!")
   if request.method == 'POST':
     session.delete(genre)
     session.commit()
@@ -630,6 +641,8 @@ def edit_author(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   author = session.query(Author).filter_by(id = id).one()
+  if author.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit authors that you created. <a href='/catalog/author/create'>Create your own author</a> if you're up for it!")
   if request.method == 'POST':
     if request.form['first_name']:
       author.first_name = request.form['first_name']
@@ -660,6 +673,8 @@ def delete_author(id):
   if 'username' not in login_session:
     return redirect(url_for('login'))
   author = session.query(Author).filter_by(id = id).one()
+  if author.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit authors that you created. <a href='/catalog/author/create'>Create your own author</a> if you're up for it!")
   if request.method == 'POST':
     session.delete(author)
     session.commit()
@@ -675,6 +690,8 @@ def remove_author(book_id, author_id):
     return redirect(url_for('login'))
   book = session.query(Book).filter_by(id = book_id).one()
   author = session.query(Author).filter_by(id = author_id).one()
+  if book.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit books that you created. <a href='/catalog/book/create'>Create your own book</a> if you're up for it!")
   if author in book.authors:
     try:
       print 'Removing {} from {}!'.format(author.last_name, book.title)
@@ -694,6 +711,8 @@ def remove_genre(book_id, genre_id):
     return redirect(url_for('login'))
   book = session.query(Book).filter_by(id = book_id).one()
   genre = session.query(Genre).filter_by(id = genre_id).one()
+  if book.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit books that you created. <a href='/catalog/book/create'>Create your own book</a> if you're up for it!")
   if genre in book.genres:
     try:
       print 'Removing {} from {}!'.format(genre.type, book.title)
@@ -713,6 +732,8 @@ def author_remove_book(author_id, book_id):
     return redirect(url_for('login'))
   author = session.query(Author).filter_by(id = author_id).one()
   book = session.query(Book).filter_by(id = book_id).one()
+  if author.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit authors that you created. <a href='/catalog/author/create'>Create your own author</a> if you're up for it!")
   if book in author.books:
     try:
       print 'Removing {} from {} list!'.format(book.title, author.last_name)
@@ -732,6 +753,8 @@ def genre_remove_book(genre_id, book_id):
     return redirect(url_for('login'))
   genre = session.query(Genre).filter_by(id = genre_id).one()
   book = session.query(Book).filter_by(id = book_id).one()
+  if genre.user_id != login_session['user_id']:
+    flash("Sorry, you can only edit genres that you created. <a href='/catalog/genre/create'>Create your own genre</a> if you're up for it!")
   if book in genre.books:
     try:
       print 'Removing {} from {}!'.format(book.title, genre.type)
